Add uthash support for resend.c
The idea of a "priority queue" is expressed elsewhere in the code,
but it is not fully realized in the linked list implementation.

This exposes the idea to other parts of babel.

The second thing, not exposed, is a switch to the uthash
( http://troydhanson.github.io/uthash/ ) library for lookups
which automagically makes available both a jenkins hash of the
34 byte tuple, but has in-place deletion, and sorting capabilities.

The difference in performance was amazing with lots of routes.